+++ changelog.md
@@ -1,4 +1,11 @@
 ------------------------------------------------------
+Version 1.1.2
+------------------------------------------------------
+- commands now work with command blocks
+- fixed wrong command usage
+- added parameters to the /addSpawnPoint command
+
+------------------------------------------------------
 Version 1.1.1
 ------------------------------------------------------
 - command localization fix